requirements.txt
bandit==1.7.6
black==24.1.1
# Core Project Dependencies
boto3>=1.35.0
duckdb==1.1.3
flake8==7.0.0
flake8-bugbear==24.1.17
flake8-docstrings==1.7.0
fsspec>=2024.2.0
ibis-framework[duckdb]==9.5.0
isort==5.13.2
mypy==1.8.0

# Type Checking
mypy-extensions==1.0.0
pandas<3.0.0
pep8-naming==0.13.3

# Development and Pre-Commit Dependencies
pre-commit==3.6.2
pyarrow<13,>=12.0.1

# Additional Linting and Code Quality
pycodestyle==2.11.1
pyflakes==3.2.0
requests>=2.31.0
ruff>=0.3.0
s3fs>=2024.2.0
sqlfluff==3.0.0
sqlglot~=25.20.2
sqlmesh==0.122.3
types-freezegun
types-python-dateutil
types-PyYAML
types-requests
